refactor(persistence,plugin-runtime): move the drain's progress file to where both readers can reach it
Two surfaces read this file and they sit on opposite sides of a package wall:
the CLI's status block, and the sync panel on the dashboard, which may not
import the plugin runtime at all. The module only ever depended on persistence,
and persistence already owned its filename constant for the same reason, so the
whole module moves rather than being split.

The whole module, not just the reader: the reader and the writer share a
SPEC_VERSION, and a validator separated from the writer it validates drifts on
the next bump. The old path stays as a re-export, so every existing caller is
untouched.

Two pinned-set guards fired on the way, both correctly: the public symbol set
and the list of suites that build a temp tree without opening a store. Each is
updated deliberately, the second with the reason the list requires -- this
module reads and writes one JSON file and never opens the store.
